only rsync if folder is present

+++ install_on_csb.sh
@@ -6,8 +6,10 @@ sed -i 's/csb_hostname.*/csb_hostname = moriarty.orion/' /etc/ansible/facts.d/cs
 sed -i 's/^dhill.*/moriarty.orion ansible_connection=local/' /etc/ansible/hosts
 
 yum install -y terminator vim snap shairport-sync vlc thunderbird net-snmp google-chrome-unstable gnome-classic-session git-lfs hexchat
-rsync -avgo --remove-source-files /home/dhill_restore/ /home/dhill/
-find /home/dhill_restore/ -depth -type d -empty -delete
+if [ -d /home/dhill_restore ]; then
+  rsync -avgo --remove-source-files /home/dhill_restore/ /home/dhill/
+  find /home/dhill_restore/ -depth -type d -empty -delete
+fi
 
 cp etc/gdm/* /etc/gdm/
 cp etc/snmp/* /etc/snmp/
